added tests for bmi calculation

+++ bmi_calculator_app/lib/bloc/calculator_event.dart
@@ -10,14 +10,13 @@ class CalculatorReset extends CalculatorEvent {}
 class CalculatorHeightChanged extends CalculatorEvent {
   const CalculatorHeightChanged({this.height});
 
-  final int height;
+  final double height;
 }
 
-// sorry: we use integers: if one is 69.5 kg it's 69 or 70, nothing between
 class CalculatorWeightChanged extends CalculatorEvent {
   const CalculatorWeightChanged({this.weight});
 
-  final int weight;
+  final double weight;
 }
